requirements.txt
fastapi
uvicorn
azure-servicebus
pydantic
python-dotenv